Added new challenge to speed room
