--- Assets/Scenes/SpeedRooms/Spikes.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ebb616b63a41cd646a90400fa03889ca
