--- Assets/Scripts/Spawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4aef6b25ea5406649a18301e210b9b82
